default __marin__ implementation for base class ConstitutiveLaw
Implementation of default __marin__ function for base class ConstitutiveLaw. In this way each new constitutive law has a free function for getting marin coefficients.
This is achieved by discretization of constitutive law a a piecewise linear function.
For the discretized constitutive law the marin coefficients are returned.

Included tests for Sargin constitutive law and for section response with sargin law (the marin integration uses the newly defined base __marin__ function)