fix: Constitutive law fixes and refactoring (#202)
* included isscalar check on constitutive laws

Constitutive laws took as input an ArrayLike. This was converted to a numpy array and processed subsequently. Now if the user input a scalar, a scalar is returned.

* Refactored constitutive_laws

Refactored constitutive laws created a sub-module. This will permit in the future (or even now) to split constitutive laws in separate files making it more easy to be maintained.

* fix typing & split files

Fixed return types for `get_stress` and `get_tangent` methods.
Split constitutive laws into separate files.

* fixed input typing for get_stress and get_tangent