fix: bugfix moment curvature (#224)
* add tests failing with current implementation

* more robust prefind_range_curvature_equilibrium

Prefind range curvature equilibrium is a simple algorithm for quickly finding a range where to apply bisection algorithm.
In the original implementation there was a check needed for cases where the prefind is diverging from solution (this can occur when the section is close to failure). In some circumstances, due to oscillations in the function delta_N vs. curvature, the algorithm could interpret a diverging even if not real.
For this reason the function has been improved with a more robust version.
The relevant tests added with commit https://github.com/fib-international/structuralcodes/pull/224/commits/24b0501168bbee78aef5429e61f11b10a11c2cc2 now passes.

* interrupt algorithm with max number of restarts